chore: fix test rename follow-ups
