+++ .github/workflows/test.yml
@@ -174,8 +174,8 @@ jobs:
           - postgres
           - mysql
         test:
-          - unit
-          - integration
+          - feature
+          - web
     services:
       db-mysql:
         image: mariadb:11@sha256:fcc7fcd7114adb5d41f14d116b8aac45f94280d2babfbbb71b4782922ee6d8d4
