+++ pkg/webtests/_test.go.tpl
@@ -14,7 +14,7 @@
 //   You should have received a copy of the GNU General Public License
 //   along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package integrations
+package webtests
 
 import (
 	"code.vikunja.io/api/pkg/models"
