+++ pkg/webtests/token_test.go
@@ -14,31 +14,21 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package integrations
+package webtests
 
 import (
 	"net/http"
 	"testing"
 
 	apiv1 "code.vikunja.io/api/pkg/routes/api/v1"
-
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-func TestUserProject(t *testing.T) {
+func TestCheckToken(t *testing.T) {
 	t.Run("Normal test", func(t *testing.T) {
-		rec, err := newTestRequestWithUser(t, http.MethodPost, apiv1.UserList, &testuser1, "", nil, nil)
+		rec, err := newTestRequestWithUser(t, http.MethodPost, apiv1.CheckToken, &testuser1, "", nil, nil)
 		require.NoError(t, err)
-		assert.Equal(t, "null\n", rec.Body.String())
-	})
-	t.Run("Search for user3", func(t *testing.T) {
-		rec, err := newTestRequestWithUser(t, http.MethodPost, apiv1.UserList, &testuser1, "", map[string][]string{"s": {"user3"}}, nil)
-		require.NoError(t, err)
-		assert.Contains(t, rec.Body.String(), `user3`)
-		assert.NotContains(t, rec.Body.String(), `user1`)
-		assert.NotContains(t, rec.Body.String(), `user2`)
-		assert.NotContains(t, rec.Body.String(), `user4`)
-		assert.NotContains(t, rec.Body.String(), `user5`)
+		assert.Contains(t, rec.Body.String(), `🍵`)
 	})
 }
