+++ pkg/webtests/user_password_request_token_test.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-package integrations
+package webtests
 
 import (
 	"net/http"
@@ -22,32 +22,30 @@ import (
 
 	apiv1 "code.vikunja.io/api/pkg/routes/api/v1"
 	"code.vikunja.io/api/pkg/user"
-
 	"github.com/labstack/echo/v4"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-func TestUserConfirmEmail(t *testing.T) {
-	t.Run("Normal test", func(t *testing.T) {
-		rec, err := newTestRequest(t, http.MethodPost, apiv1.UserConfirmEmail, `{"token": "tiepiQueed8ahc7zeeFe1eveiy4Ein8osooxegiephauph2Ael"}`, nil, nil)
+func TestUserRequestResetPasswordToken(t *testing.T) {
+	t.Run("Normal requesting a password reset token", func(t *testing.T) {
+		rec, err := newTestRequest(t, http.MethodPost, apiv1.UserRequestResetPasswordToken, `{"email": "user1@example.com"}`, nil, nil)
 		require.NoError(t, err)
-		assert.Contains(t, rec.Body.String(), `The email was confirmed successfully.`)
+		assert.Contains(t, rec.Body.String(), `Token was sent.`)
 	})
 	t.Run("Empty payload", func(t *testing.T) {
-		_, err := newTestRequest(t, http.MethodPost, apiv1.UserConfirmEmail, `{}`, nil, nil)
+		_, err := newTestRequest(t, http.MethodPost, apiv1.UserRequestResetPasswordToken, `{}`, nil, nil)
 		require.Error(t, err)
-		assert.Equal(t, http.StatusPreconditionFailed, err.(*echo.HTTPError).Code)
-		assertHandlerErrorCode(t, err, user.ErrCodeInvalidEmailConfirmToken)
+		assertHandlerErrorCode(t, err, user.ErrCodeNoUsernamePassword)
 	})
-	t.Run("Empty token", func(t *testing.T) {
-		_, err := newTestRequest(t, http.MethodPost, apiv1.UserConfirmEmail, `{"token": ""}`, nil, nil)
+	t.Run("Invalid email address", func(t *testing.T) {
+		_, err := newTestRequest(t, http.MethodPost, apiv1.UserRequestResetPasswordToken, `{"email": "user1example.com"}`, nil, nil)
 		require.Error(t, err)
-		assertHandlerErrorCode(t, err, user.ErrCodeInvalidEmailConfirmToken)
+		assert.Equal(t, http.StatusBadRequest, err.(*echo.HTTPError).Code)
 	})
-	t.Run("Invalid token", func(t *testing.T) {
-		_, err := newTestRequest(t, http.MethodPost, apiv1.UserConfirmEmail, `{"token": "invalidToken"}`, nil, nil)
+	t.Run("No user with that email address", func(t *testing.T) {
+		_, err := newTestRequest(t, http.MethodPost, apiv1.UserRequestResetPasswordToken, `{"email": "user1000@example.com"}`, nil, nil)
 		require.Error(t, err)
-		assertHandlerErrorCode(t, err, user.ErrCodeInvalidEmailConfirmToken)
+		assertHandlerErrorCode(t, err, user.ErrCodeUserDoesNotExist)
 	})
 }
